requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.7
aiosignal==1.3.2
attrs==25.3.0
audioop-lts==0.2.1
cffi==1.17.1
discord.py==2.5.2
frozenlist==1.6.2
idna==3.10
multidict==6.4.4
propcache==0.3.1
pycparser==2.22
PyNaCl==1.5.0
yarl==1.20.0
yt-dlp==2025.5.22
